+++ pymatviz/utils/__init__.py
@@ -1,0 +1,7 @@
+from __future__ import annotations
+
+import os
+
+PKG_DIR = os.path.dirname(__file__)
+ROOT = os.path.dirname(PKG_DIR)
+
