+++ pymatviz/utils/testing.py
@@ -1,0 +1,5 @@
+from __future__ import annotations
+
+from pymatviz.utils import ROOT
+
+TEST_FILES: str = f"{ROOT}/tests/files"
